move playmode components from lib to route
