+++ src/routes/listbuilder/components/ListbuilderComponents/FormationMenu.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 	import { toastController } from "$lib/stores";
 	import { List, type ListFormation } from "$lib/types/list.svelte";
-	import { sendListToPlay } from "$lib/playmode/utilities/playmodeUtilities";
+	import { sendListToPlay } from "$lib/playmode/playmode";
 	import { exportToJeff } from "../../utilities/export.svelte";
 	import FindUnitAvailabilityModal from "../modals/FindUnitAvailabilityModal.svelte";
 	import { DropdownMenu } from "$lib/generic";
