+++ src/routes/play/components/CritBoxMech.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import type { PlayUnit } from "$lib/playmode/types";
-	import type { CritList } from "$lib/playmode/utilities/playmodeAutomation";
+	import type { CritList } from "../utilities/playmodeAutomation";
 
 	type Props = {
 		unit: PlayUnit;
