+++ src/routes/play/components/CritBoxProto.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import type { PlayUnit } from "$lib/playmode/types";
-	import type { CritList } from "$lib/playmode/utilities/playmodeAutomation";
+	import type { CritList } from "../utilities/playmodeAutomation";
 
 	type Props = {
 		unit: PlayUnit;
@@ -11,18 +11,6 @@
 </script>
 
 <div class="crit-block-body">
-	<p class="crit-header">Engine</p>
-	<div class="crit-line">
-		{#each { length: 1 }, index}
-			<div
-				class="pip"
-				class:damaged={critCount.current.engine > index}
-				class:pending-pip={critCount.current.engine <= index && critCount.current.engine + critCount.pending.engine > index}
-			></div>
-		{/each}
-		<p>1/2 THR</p>
-	</div>
-
 	<p class="crit-header">Fire Control</p>
 	<div class="crit-line">
 		{#each { length: 4 }, index}
@@ -34,6 +22,13 @@
 		{/each}
 		<p>+2 To-Hit</p>
 	</div>
+	<p class="crit-header">MP</p>
+	<div class="crit-line">
+		{#each { length: 4 }, index}
+			<div class="pip" class:damaged={critCount.current.mp > index} class:pending-pip={critCount.current.mp <= index && critCount.current.mp + critCount.pending.mp > index}></div>
+		{/each}
+		<p>1/2 MV</p>
+	</div>
 	<p class="crit-header">Weapons</p>
 	<div class="crit-line">
 		{#each { length: 4 }, index}
@@ -43,7 +38,7 @@
 				class:pending-pip={critCount.current.weapon <= index && critCount.current.weapon + critCount.pending.weapon > index}
 			></div>
 		{/each}
-		<p>-1 Dmg</p>
+		<p>-1 Damage</p>
 	</div>
 </div>
 
@@ -51,6 +46,7 @@
 	p {
 		color: black;
 	}
+
 	.crit-block-body {
 		display: grid;
 		grid-template-columns: min-content 1fr;
@@ -68,7 +64,7 @@
 		display: flex;
 		gap: 0.5cqw;
 		align-items: center;
-		p {
+		& p {
 			font-size: 2.25cqmax;
 			margin-left: 0.5cqw;
 			white-space: nowrap;
