+++ src/routes/play/components/CritModal.svelte
