+++ src/routes/play/components/DisplayOptionsPopover.svelte
