+++ src/routes/play/components/HeatModal.svelte
