+++ src/routes/play/components/SpecialModal.svelte
