+++ src/routes/play/schema/playmode.ts
